Added console manager (#4)
