+++ console_manager.py
@@ -1,0 +1,14 @@
+import os
+
+
+def press_enter():
+    input("Press Enter key to continue... ")
+
+
+def clear():
+    if os.name == 'nt':
+        console_command = 'cls'
+    else:
+        console_command = 'clear'
+
+    os.system(console_command)
